refactor(link-button-component): make it a common component, with label as a prop
